chore(infrastructure): create infrastructure for version 1.x
