+++ package.json
@@ -1,36 +1,49 @@
 {
   "name": "nDaterangepicker",
-  "version": "0.1.10",
-  "description": "Angular.js wrapper for dangrossman/bootstrap-daterangepicker",
+  "version": "1.0.0",
+  "description": "Angular 1.x self-sufficient datepicker and daterangepicker solution",
+  "author": {
+    "name": "Eugene Serkin",
+    "email": "jeserkin@gmail.com",
+    "url": "blog.art-coder.com"
+  },
+  "license": {
+    "type": "MIT",
+    "url": "http://www.opensource.org/licenses/mit-license.php"
+  },
   "repository": {
     "type": "git",
     "url": "git://github.com/jeserkin/nDaterangepicker.git"
   },
-  "dependencies": {},
-  "devDependencies": {
-    "chai": "^1.9.1",
-    "chai-jquery": "^1.2.3",
-    "grunt": "^0.4.5",
-    "grunt-wiredep": "^2.0.0",
-    "grunt-contrib-concat": "^0.4.0",
-    "grunt-contrib-uglify": "^0.4.0",
-    "grunt-ng-annotate": "^0.9.2",
-    "grunt-karma": "^0.9.0",
-    "load-grunt-tasks": "^1.0.0",
-    "karma": "^0.12.28",
-    "karma-chai": "^0.1.0",
-    "karma-jasmine": "^0.3.5",
-    "karma-mocha": "^0.1.8",
-    "karma-mocha-reporter": "^1.0.2",
-    "karma-phantomjs-launcher": "^0.1.4",
-    "karma-ng-html2js-preprocessor": "^0.1.2",
-    "karma-ng-scenario": "^0.1.0",
-    "mocha": "^1.21.4"
-  },
-  "engines": {
-    "node": ">=0.10.0"
-  },
   "scripts": {
-    "test": "grunt karma"
+    "build": "gulp",
+    "test": "gulp karma"
+  },
+  "dependencies": {
+    "angular": "1.5.8",
+    "moment": "2.15.1"
+  },
+  "devDependencies": {
+    "gulp": "3.9.1",
+    "gulp-util": "3.0.7",
+    "load-gulp-tasks": "0.1.0",
+    "gulp-load-plugins": "1.3.0",
+    "gulp-concat": "2.6.0",
+    "gulp-header": "1.8.8",
+    "gulp-ng-annotate": "2.0.0",
+    "gulp-angular-templatecache": "2.0.0",
+    "gulp-uglify": "2.0.0",
+    "wiredep": "4.0.0",
+
+    "karma": "1.3.0",
+    "karma-chai": "0.1.0",
+    "karma-mocha": "^0.1.8",
+    "karma-mocha-reporter": "2.2.0",
+    "karma-phantomjs-launcher": "1.0.2",
+    "karma-ng-html2js-preprocessor": "1.0.0",
+    "karma-ng-scenario": "1.0.0",
+    "mocha": "3.1.0",
+    "chai": "3.5.0",
+    "chai-jquery": "2.0.0"
   }
 }
