+++ src/index.js
@@ -1,0 +1,5 @@
+(function(angular) {
+  'use strict';
+
+  angular.module('nDaterangepicker', ['nDaterangepickerTemplates']);
+})(angular);
